requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
Authlib==1.3.1
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.149
botocore==1.34.149
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
flask==3.0.3
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib-metadata==8.2.0
importlib-resources==6.4.0
isodate==0.6.1
itsdangerous==2.2.0
jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pkgutil-resolve-name==1.3.10
protobuf==5.27.2
pyarrow==17.0.0
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
pydeck==0.9.1
pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
streamlit==1.37.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typer==0.12.3
typing-extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
werkzeug==3.0.3
zipp==3.19.2